[proliferate] split automation cloud executor stages
